API

MonotoneSplines.Spl — Type

This is a Spl

Fields

H: an RObject generated by splines::bs()

β: the coefficients for the B-spline.

MonotoneSplines.build_model — Method

build_model(x::AbstractVector{T}, J::Int)
build_model(x::AbstractVector{T}, scaled::Bool)
build_model(x::AbstractVector{T})

Returns

B-spline design matrix B at x for cubic splines

Bnew at new points xnew if it is not nothing

L = nothing: keep same return list for model of smoothing splines

J: number of basis functions, which does not change for cubic splines, so it is only intended for smoothing splines

mx, rx, idx, idx0: the remaining only for smoothing splines

MonotoneSplines.conf_band_width — Method

conf_band_width(CIs::AbstractMatrix)

Calculate width of confidence bands.

MonotoneSplines.coverage_prob — Method

coverage_prob(CIs::AbstractMatrix, y0::AbstractVector)

Calculate coverage probability given n x 2 CI matrix CIs and true vector y0 of size n.

MonotoneSplines.cv_mono_ss — Method

cv_mono_ss(x::AbstractVector{T}, y::AbstractVector{T}, λs::AbstractVector{T})

Cross-validation for monotone fitting with smoothing spline on y ~ x among parameters λs.

MonotoneSplines.div_into_folds — Method

div_into_folds(N::Int; K = 10, seed = 1234)

Equally divide 1:N into K folds with random seed seed. If seed is negative, it is a non-random division, where the i-th fold would be the i-th equidistant range.

MonotoneSplines.eval_penalty — Method

eval_penalty(model::Spl{T}, x::AbstractVector{T})

Evaluate the penalty matrix by R's fda::eval.penalty. To make sure the corresponding design matrix contructed by fda::eval.basis is the same as model.H, it asserts the norm difference should be smaller than sqrt(eps()).

MonotoneSplines.fit — Function

fit(X, y, paras, method)

paras is either the number of basis functions, or the sequence of interior knots. Return a Spl object.

n = 100
x = rand(n) * 2 .- 1
y = x .^3 + randn(n) * 0.01
res = fit(x, y, 10, "monotone")

MonotoneSplines.gen_data — Method

gen_data(n, σ, f::Union{Function, String}; xmin = -1, xmax = 1, k = 10)

Generate n data points (xi, yi) from curve f with noise level σ, i.e., yi = f(xi) + N(0, σ^2).

It returns four vectors, x, y, x0, y0, where

x, y: pair points of length n.

x0, y0: true curve without noise, represented by k*n points.

MonotoneSplines.mono_ss — Function

mono_ss(B::AbstractMatrix, y::AbstractVector, L::AbstractMatrix, J::Int, λ::AbstractFloat)

Monotone Fitting with Smoothing Splines given design matrix B and cholesky-decomposed matrix L.

MonotoneSplines.predict — Method

predict(model::Spl{T}, xs::AbstractVector{T})
predict(X::Vector{Float64}, y::Vector{Float64}, J::Int, Xnew::AbstractVector{Float64}, ynew::AbstractVector{Float64}
predict(X::Vector{Float64}, y::Vector{Float64}, J::Int, Xnew::Vector{Float64}, ynew::Vector{Float64}, σ::Vector{Float64}

Make prediction based on fitted Spl on new points xs. If Xnew is provided, then also returns the prediction error ‖yhat - ynew‖_2^2.